smart-question/requirements.txt
fastapi
uvicorn
python-multipart
starlette
pdfplumber
faiss-cpu
numpy
dashscope
python-docx
typing-extensions
